[#174] refactor: 공통 유틸로 사용 될 수 있는 함수 분리
PR #189 에서 받은 피드백 적용
시간 포메팅 함수, 비어있는 ARRAY 판별 함수 공통 유틸로 분리
