+++ src/utils/isDataEmpty.ts
@@ -1,0 +1,3 @@
+export default function isDataEmpty(data: unknown[]) {
+  return data.length === 0;
+}
